Rename files + update license
