+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 - Add support for MermaidJS. See https://github.com/haidaraM/ansible-playbook-grapher/issues/137
 - Update various Dependencies: pytest, pytest-cov, ansible-core, pyquery etc...
 - ci: Add dependabot for github-actions
-- Other internal fixes...
+- Rename some tests files
+- ...
 
 ## Breaking changes
 
@@ -14,6 +15,7 @@ as a library inside another project:
 
 - Completely refactor the rendering part of the part by making it more extensible in order to support Mermaid. 
 - Fill the plays, blocks and node with color to make them more visible in the output
+- Rename the file `graph.py` to `graph_model.py`
 - Use the concatenation of the playbook names as the output filename when graphing multiple playbooks instead of the
   first playbook.
